+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-...
+- The fork names are auto-completed for `sync` (`--project-fork` option) and `bc` (`--source-project-fork` and `--target-project-fork` options) commands.
 
 ### Changed
 ...